Changes to the constructor funct.
