--- Limbs.txt
@@ -1,179 +1,0 @@
-# This is the transformer limb manipulation file, you may change any value but do NOT change the layout of the file. If you do, STOK might not generate properly.
-# Under here are the common values to all limbs
-# Number of limbs
-8
-
-# Limb colour
-(0,0,0,1)
-
-# Distance between outer containment radius and the limb
-1555.0
-
-# Default sphere radius
-500.0
-
-# Offset for (port1-port2)/2 [degrees] (yes or no)
-"yes"
-
-####################################################################
-# And under here is everything else
-# Every new limb is to be set identically (as in the same format) as the first one
-# 1:
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-"Custom limb name"
-
-# 2: ##################
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 3: ##################
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 4: ##################
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 5: ##################
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 6: ##################
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 7: ##################
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 8: ##################
-# Limb length
-2000.0
-
-# Limb width
-1000.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-12000.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 9: ##################
-# Limb length
-10.0
-
-# Limb width
-8.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-0.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
-
-# 10: #################
-# Limb length
-10.0
-
-# Limb width
-8.0
-
-# Limb radius (if left 0 the radius is the default value set in the common values section)
-0.0
-
-# Limb height (if left 0 it defaults to the containment height)
-0.0
-
-# Limb name (optional if left 0 it will default to generic limb naming)
-0
