--- Solenoid.txt
@@ -1,12 +1,0 @@
-# This is the solenoid manipulation file, you may change the values but do NOT change the layout of the file
-# Solenoid colour the format is as follows (red, green, blue, transparency) valid values (0-1)
-(1,0,0,1)
-
-# Solenoid radius (this will also determine the inner radius of the containment)
-900.0
-
-# Height of solenoid
-6400.0
-
-# Bounding box wall thickness
-2000.0
