+++ test.py
@@ -1,0 +1,5 @@
+"""_summary_"""
+import cadquery as cq
+from stok_modules import STOK
+
+show_object(STOK().transformer_limbs())
